backend/pyproject.toml
[tool.poetry]
name = "todo-test"
version = "0.1.0"
description = ""
authors = ["Kai Odi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
pydantic-settings = "^2.8.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
